block view/update/delete of discussions to not owners, block view to not owners or members

+++ app/Http/Controllers/DiscussionController.php
@@ -7,9 +7,27 @@ use App\Models\Discussion;
 
 class DiscussionController extends Controller
 {
+    /** @var JWTGuard */
+    private $auth;
+    /** @var User */
+    private $user;
+
+    public function __construct()
+    {
+        $this->auth = auth();
+        $this->user = $this->auth->user();
+    }
+    
     public function index()
     {
-        $discussions = Discussion::all();
+        $discussions = Discussion::where('is_public', true)->orWhere(
+            function ($query) {
+                $query->where('is_public', false)
+                    ->where('author_id', $this->user->id)
+                    ->orWhereHas('members', fn($query) => $query->where('user_id', $this->user->id));
+            }
+        )->get();
+
         return response()->json($discussions, 200);
     }
 
@@ -28,6 +46,8 @@ class DiscussionController extends Controller
 
     public function show(Discussion $discussion)
     {
+        $this->authorize('view', $discussion);
+
         $discussion->load('topics', 'members');
 
         return response()->json($discussion, 200);
@@ -35,6 +55,8 @@ class DiscussionController extends Controller
 
     public function update(DiscussionRequest $request, Discussion $discussion)
     {
+        $this->authorize('update', $discussion);
+
         $data = $request->validated();
 
         $members = $this->getMembersFromData($data);
@@ -48,6 +70,8 @@ class DiscussionController extends Controller
 
     public function destroy(Discussion $discussion)
     {
+        $this->authorize('delete', $discussion);
+
         $discussion->delete();
         return response()->json(['message' => 'Discussion deleted successfully'], 200);
     }
@@ -70,5 +94,26 @@ class DiscussionController extends Controller
         $discussion->load('members');
 
         return $discussion;
+    }
+
+    private function authorize(string $ability, Discussion $discussion): void
+    {
+        if ($discussion->is_public) {
+            return;
+        }
+
+        if ($discussion->author_id === $this->user->id) {
+            return;
+        }
+
+        if ($ability === 'view' && $discussion->members->contains($this->user)) {
+            return;
+        }
+
+        if ($ability === 'view') {
+            abort(404, 'Not found');
+        }
+
+        abort(403, 'Unauthorized');
     }
 }
